requirements.txt
# Core dependencies
google-genai>=0.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
ruff>=0.1.0
isort>=5.12.0

# API framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Optional: Configuration management
pyyaml>=6.0
